Layout Test http/tests/cache-storage/cache-clearing.https.html is failing https://bugs.webkit.org/show_bug.cgi?id=178200
Patch by Youenn Fablet <youenn@apple.com> on 2017-10-12
Reviewed by Chris Dumez.

Source/WebKit:

Deleting only the per-origin folders inside the CacheStorage engine root folder.

* NetworkProcess/cache/CacheStorageEngine.cpp:
(WebKit::CacheStorage::Engine::clearAllCaches):
* UIProcess/API/C/WKWebsiteDataStoreRef.cpp:
(WKWebsiteDataStoreRemoveAllFetchCaches):
(WKWebsiteDataStoreRemoveFetchCacheForOrigin):
* UIProcess/API/C/WKWebsiteDataStoreRef.h:

Tools:

Making clearDOMCache wait for removal completion to exit.
This removes the risk to interact with the cache while deleting it which will end up be racy in tests.

* WebKitTestRunner/InjectedBundle/TestRunner.cpp:
(WTR::TestRunner::clearDOMCache):
* WebKitTestRunner/TestController.cpp:
(WTR::ClearDOMCacheCallbackContext::ClearDOMCacheCallbackContext):
(WTR::clearDOMCacheCallback):
(WTR::TestController::clearDOMCache):
* WebKitTestRunner/TestInvocation.cpp:
(WTR::TestInvocation::didReceiveMessageFromInjectedBundle):
(WTR::TestInvocation::didReceiveSynchronousMessageFromInjectedBundle):

LayoutTests:

Splitting cache-clearing.https.html in two sub tests and marking them as Pass/Timeout for the moment.

* platform/win/TestExpectations: Skipping cache/sw tests.
* TestExpectations:
* http/tests/cache-storage/cache-clearing.https.html: Removed.
* http/tests/cache-storage/cache-clearing-all.https-expected.txt: Added.
* http/tests/cache-storage/cache-clearing-all.https.html: Added.
* http/tests/cache-storage/cache-clearing-origin.https-expected.txt: Added.
* http/tests/cache-storage/cache-clearing-origin.https.html: Added.

Canonical link: https://commits.webkit.org/194482@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@223265 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/NetworkProcess/cache/CacheStorageEngine.cpp
@@ -353,8 +353,11 @@ void Engine::clearAllCaches(CallbackAggregator& taskHandler)
     if (!shouldPersist())
         return;
 
-    m_ioQueue->dispatch([filename = m_rootPath.isolatedCopy(), taskHandler = makeRef(taskHandler)] {
-        deleteDirectoryRecursively(filename);
+    m_ioQueue->dispatch([path = m_rootPath.isolatedCopy(), taskHandler = makeRef(taskHandler)] {
+        for (auto& filename : WebCore::listDirectory(path, "*")) {
+            if (WebCore::fileIsDirectory(filename, WebCore::ShouldFollowSymbolicLinks::No))
+                deleteDirectoryRecursively(filename);
+        }
     });
 }
